+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - fix library switch not working upon first launch
+- fix out ouf bounds crash when current_library is higher than number of library paths
 
 ### Misc